- Рефакторинг на подход Single Activity - Подключить Jetpack Navigation Component - Реализовать паттерн навигации с использованием BottomNavigationView
