+++ app/build.gradle.kts
@@ -56,6 +56,8 @@ dependencies {
     implementation ("com.squareup.retrofit2:retrofit:2.11.0")
     implementation ("com.squareup.retrofit2:converter-gson:2.11.0")
     implementation ("io.insert-koin:koin-android:3.3.0")
-    implementation("androidx.fragment:fragment-ktx:1.5.5")
+    implementation ("androidx.navigation:navigation-fragment-ktx:2.5.3")
+    implementation ("androidx.navigation:navigation-ui-ktx:2.5.3")
+    implementation("androidx.fragment:fragment-ktx:1.5.6")
     implementation("androidx.viewpager2:viewpager2:1.0.0")
 }
