--- app/src/main/java/com/example/playlistmaker/library/ui/LibraryActivity.kt
@@ -1,39 +1,0 @@
-package com.example.playlistmaker.library.ui
-
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import com.example.playlistmaker.R
-import com.example.playlistmaker.databinding.ActivityLibraryBinding
-import com.google.android.material.tabs.TabLayoutMediator
-
-class LibraryActivity : AppCompatActivity() {
-
-    private lateinit var binding : ActivityLibraryBinding
-    private lateinit var tabMediator: TabLayoutMediator
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        binding = ActivityLibraryBinding.inflate(layoutInflater)
-        setContentView(binding.root)
-
-        binding.myToolbar.setNavigationOnClickListener() {
-            finish()
-        }
-
-        binding.viewPager.adapter = LibraryViewPagerAdapter(supportFragmentManager, lifecycle)
-
-        tabMediator = TabLayoutMediator(binding.tabLayout, binding.viewPager) { tab, position ->
-            when (position) {
-                0 -> tab.text = getString(R.string.favorites)
-                1 -> tab.text = getString(R.string.playlists)
-            }
-        }
-        tabMediator.attach()
-
-    }
-
-    override fun onDestroy() {
-        super.onDestroy()
-        tabMediator.detach()
-    }
-}
