+++ app/src/main/java/com/example/playlistmaker/library/ui/LibraryFragment.kt
@@ -1,0 +1,45 @@
+package com.example.playlistmaker.library.ui
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import com.example.playlistmaker.R
+import com.example.playlistmaker.databinding.FragmentLibraryBinding
+import com.google.android.material.tabs.TabLayoutMediator
+
+class LibraryFragment : Fragment() {
+
+    private var _binding: FragmentLibraryBinding? = null
+    private val binding get() = _binding!!
+
+    private lateinit var tabMediator: TabLayoutMediator
+
+    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
+        _binding = FragmentLibraryBinding.inflate(inflater, container, false)
+        return binding.root
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+
+        binding.viewPager.adapter = LibraryViewPagerAdapter(childFragmentManager, lifecycle)
+
+        tabMediator = TabLayoutMediator(binding.tabLayout, binding.viewPager) { tab, position ->
+            when (position) {
+                0 -> tab.text = getString(R.string.favorites)
+                1 -> tab.text = getString(R.string.playlists)
+            }
+        }
+        tabMediator.attach()
+
+    }
+
+    override fun onDestroyView() {
+        super.onDestroyView()
+        tabMediator.detach()
+        _binding = null
+    }
+
+}
