+++ app/src/main/java/com/example/playlistmaker/library/ui/LibraryViewPagerAdapter.kt
@@ -13,7 +13,6 @@ class LibraryViewPagerAdapter(fragmentManager: FragmentManager, lifecycle: Lifec
 
     override fun createFragment(position: Int): Fragment {
         return when (position) {
-            0 -> FavoritesFragment.newInstance()
             1 -> PlaylistsFragment.newInstance()
             else -> FavoritesFragment.newInstance()
         }
