--- app/src/main/java/com/example/playlistmaker/main/ui/MainActivity.kt
@@ -1,44 +1,0 @@
-package com.example.playlistmaker.main.ui
-
-import android.content.Intent
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import android.widget.Button
-import com.example.playlistmaker.R
-import com.example.playlistmaker.search.ui.SearchActivity
-import com.example.playlistmaker.settings.ui.SettingsActivity
-import com.example.playlistmaker.library.ui.LibraryActivity
-
-class MainActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
-
-
-        val buttonSearch = findViewById<Button>(R.id.search_button)
-
-        buttonSearch.setOnClickListener {
-            val searchIntent = Intent(this, SearchActivity::class.java)
-            startActivity(searchIntent)
-        }
-
-        val buttonLib = findViewById<Button>(R.id.library_button)
-
-        buttonLib.setOnClickListener {
-            val libraryIntent = Intent(this, LibraryActivity::class.java)
-            startActivity(libraryIntent)
-        }
-
-
-        val buttonSet = findViewById<Button>(R.id.settings_button)
-
-        buttonSet.setOnClickListener {
-            val settingsIntent = Intent(this, SettingsActivity::class.java)
-            startActivity(settingsIntent)
-        }
-
-
-
-
-    }
-}
