+++ app/src/main/res/color/bottom_menu.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <item android:state_checked="true" android:color="@color/bottom_menu_active" />
+    <item android:color="@color/bottom_menu_default"  />
+
+</selector>
