--- app/src/main/res/layout/activity_main.xml
@@ -1,75 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical"
-    style="@style/MainActivityStyle">
-
-    <FrameLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content">
-
-        <FrameLayout
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:padding="@dimen/value_4">
-
-            <FrameLayout
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:paddingStart="@dimen/value_12"
-                android:paddingEnd="@dimen/value_12"
-                android:paddingTop="@dimen/value_10"
-                android:paddingBottom="@dimen/value_12">
-
-
-                <TextView
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:text="@string/app_name"
-                    android:textColor="@color/white"
-                    android:textSize="@dimen/big_size" />
-
-            </FrameLayout>
-
-        </FrameLayout>
-
-    </FrameLayout>
-
-
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:layout_marginTop="@dimen/value_24"
-        android:layout_marginHorizontal="@dimen/value_16"
-        android:layout_marginBottom="@dimen/value_28"
-        android:orientation="vertical"
-        >
-
-
-        <Button
-            android:id="@+id/search_button"
-            style="@style/MainActivityButtonStyle"
-            android:text="@string/search"
-            app:icon="@drawable/search"/>
-
-        <Button
-            android:id="@+id/library_button"
-            style="@style/MainActivityButtonStyle"
-            android:text="@string/library"
-            app:icon="@drawable/library"/>
-
-        <Button
-            android:id="@+id/settings_button"
-            style="@style/MainActivityButtonStyle"
-            android:layout_marginBottom="@dimen/value_0"
-            android:text="@string/settings"
-            app:icon="@drawable/settings"/>
-
-
-
-    </LinearLayout>
-
-</LinearLayout>
