+++ app/src/main/res/layout/activity_player.xml
@@ -4,7 +4,8 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
-    style="@style/OtherActivityStyle">
+    style="@style/OtherActivityStyle"
+    tools:context=".player.ui.Player">
 
     <androidx.constraintlayout.widget.Guideline
         android:id="@+id/left_guideline"
