+++ app/src/main/res/layout/fragment_library.xml
@@ -2,7 +2,10 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:orientation="vertical">
+    xmlns:tools="http://schemas.android.com/tools"
+    android:orientation="vertical"
+    tools:context=".library.ui.LibraryFragment"
+    style="@style/OtherActivityStyle">
 
     <com.google.android.material.appbar.MaterialToolbar
         android:id="@+id/my_toolbar"
@@ -14,7 +17,7 @@
         android:textSize="@dimen/big_size"
         android:textFontWeight="500"
         android:fontFamily="@font/ys_display_medium"
-        app:navigationIcon="@drawable/arrow_back"
+        android:paddingStart="@dimen/value_16"
         style="@style/MyNavigationStyle"/>
