+++ app/src/main/res/layout/fragment_search.xml
@@ -7,7 +7,8 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    style="@style/OtherActivityStyle">
+    style="@style/OtherActivityStyle"
+    tools:context=".search.ui.SearchFragment">
 
     <com.google.android.material.appbar.MaterialToolbar
         android:id="@+id/my_toolbar"
@@ -19,7 +20,7 @@
         android:textSize="@dimen/big_size"
         android:textFontWeight="500"
         android:fontFamily="@font/ys_display_medium"
-        app:navigationIcon="@drawable/arrow_back"
+        android:paddingStart="@dimen/value_16"
         style="@style/MyNavigationStyle"/>
 
     <FrameLayout
