+++ app/src/main/res/layout/fragment_settings.xml
@@ -2,11 +2,13 @@
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
 
     xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
 
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    style="@style/OtherActivityStyle">
+    style="@style/OtherActivityStyle"
+    tools:context=".settings.ui.SettingsFragment">
 
     <com.google.android.material.appbar.MaterialToolbar
         android:id="@+id/my_toolbar"
@@ -18,7 +20,7 @@
         android:textSize="@dimen/big_size"
         android:textFontWeight="500"
         android:fontFamily="@font/ys_display_medium"
-        app:navigationIcon="@drawable/arrow_back"
+        android:paddingStart="@dimen/value_16"
         style="@style/MyNavigationStyle"/>
