+++ app/src/main/res/menu/bottom_navigation_menu.xml
@@ -1,0 +1,22 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <item
+        android:id="@+id/searchFragment"
+        android:icon="@drawable/search"
+        android:title="@string/search"
+        />
+
+    <item
+        android:id="@+id/libraryFragment"
+        android:icon="@drawable/library"
+        android:title="@string/library"
+        />
+
+    <item
+        android:id="@+id/settingsFragment"
+        android:icon="@drawable/settings"
+        android:title="@string/settings"
+        />
+
+</menu>
