+++ app/src/main/res/values-night/colors.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+
+    <color name="bottom_menu_default">#FFFFFFFF</color>
+
+
+</resources>
