+++ app/src/main/res/values-night/style.xml
@@ -219,4 +219,13 @@
         <item name="tabSelectedTextColor">@color/white</item>
     </style>
 
+    <style name="MyBottomMenu" parent="Widget.MaterialComponents.BottomNavigationView">
+
+        <item name="android:textSize">@dimen/menu_size</item>
+        <item name="android:textFontWeight">400</item>
+        <item name="android:fontFamily">@font/ys_display_regular</item>
+        <item name="android:background">@color/dark</item>
+
+    </style>
+
 </resources>
