+++ app/src/main/res/values/styles.xml
@@ -270,4 +270,16 @@
 
     </style>
 
+    <style name="MyBottomMenu" parent="Widget.MaterialComponents.BottomNavigationView">
+
+        <item name="android:textSize">@dimen/menu_size</item>
+        <item name="android:textFontWeight">400</item>
+        <item name="android:fontFamily">@font/ys_display_regular</item>
+        <item name="android:background">@color/white</item>
+
+    </style>
+
+
+
+
 </resources>
